requirements.test.txt
pre-commit==4.2.0
pytest==8.3.5
pytest-cov==6.0.0
pytest-homeassistant-custom-component==0.13.236
syrupy==4.8.1
